getContestStatus
