+++ libs/evm-protocols/src/common-protocol/utils.ts
@@ -299,11 +299,3 @@ export function mapToAbiRes<
     }, {} as any);
   return output;
 }
-
-export function mustBeProtocolChainId(
-  ethChainId: number,
-): asserts ethChainId is ValidChains {
-  if (!Object.values(ValidChains).includes(ethChainId)) {
-    throw new Error(`${ethChainId} is not a valid protocol eth chain id`);
-  }
-}
